Fix employee highlight persistence bug
- Fixed right-click highlight not saving to database
- Fixed highlights disappearing after date navigation
- Changed SET_HIGHLIGHTED_EMPLOYEES to MERGE instead of REPLACE
- Moved toggleHighlight to SchedulingLayout for shared access
- Updated component chain: SchedulingLayout > Scheduling/Trades > ProjectBox/UnionBox
- Highlights now persist across page refresh and date changes
